feat: add to cart spinner
